Alteração do main para adicionar porta e refactor model
