+++ financialassets/repositories/repository.go
@@ -1,6 +1,6 @@
 package repositories
 
-import "github.com/ferkze/backend-test/financialassets/domain/model"
+import "github.com/ferkze/backend-test/financialassets/model"
 
 // FinancialAssetRepository interface/contrato de repositorio de ativos financeiros
 type FinancialAssetRepository interface {
