+++ financialassets/usecases/usecases.go
@@ -1,7 +1,7 @@
 package usecases
 
 import (
-	"github.com/ferkze/backend-test/financialassets/domain/model"
+	"github.com/ferkze/backend-test/financialassets/model"
 )
 
 // FinancialAssetsUsecases casos de uso do serviço
